Translate submit btn
