+++ src/modules/translate/components/TranslateFirstLanguageTextarea.tsx
@@ -1,11 +1,10 @@
-import { useContext, memo } from "react";
-import { TbLanguage, TbX } from "react-icons/tb";
+import { memo } from "react";
+import { TbX } from "react-icons/tb";
+
 import NextButton from "@/common/components/NextButton";
 import NextTextarea from "@/common/components/NextTextarea";
-import { TranslateContext } from "./TranslateContext";
 import { ITranslateReducerAction } from "../lib/translateReducer";
-import Cookies from "js-cookie";
-import { AppContext } from "@/modules/app/components/AppContext";
+import TranslateSubmitBtn from "./TranslateSubmitBtn";
 
 interface TranslateFirstLanguageTextareaProps {
   firstLanguageText: string;
@@ -16,16 +15,6 @@ const TranslateFirstLanguageTextarea = (
   props: TranslateFirstLanguageTextareaProps
 ) => {
   const { firstLanguageText, translateDispatch } = props;
-  const { appDispatch } = useContext(AppContext)
-
-  const handleClick = async () => {
-    const token = Cookies.get('token');
-
-    if (!token) {
-      appDispatch({ key: "showLoginModal", value: true });
-      return;
-    }
-  }
 
   return (
     <div className="relative mb-4">
@@ -45,13 +34,7 @@ const TranslateFirstLanguageTextarea = (
           translateDispatch({ key: "firstLanguageText", value: e.target.value })
         }
       />
-      <NextButton
-        onClick={handleClick}
-        className="absolute bottom-4 right-2 lg:right-3 p-2"
-      >
-        <TbLanguage />
-        <span>Translate</span>
-      </NextButton>
+      <TranslateSubmitBtn />
     </div>
   );
 };
